pyproject.toml
[project]
name = "steps-on-strava"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "garminconnect>=0.2.38",
    "python-dotenv>=1.2.1",
]
